faceblur/face_detection_service/utils.py
import yaml


def read_config(config_path):

    with open("detection_config.yaml", 'r') as stream:
        try:
            config = yaml.safe_load(stream)
        except yaml.YAMLError as exc:
            print(exc)

    return config
