faceblur/face_detection_service/requirements.txt
PyYAML
celery[redis]
numpy
opencv-python-headless